abstract away providers/findPrimaryType
