+++ src/helpers/providers/web3.ts
@@ -17,7 +17,7 @@ import type { PromiEvent, provider, AbstractProvider } from 'web3-core';
 import type { JsonRpcResponse } from 'web3-core-helpers';
 import { assert } from 'ts-essentials';
 import { assertWeb3ContractHasMethods } from '../misc';
-import type { TypedDataField } from '@ethersproject/abstract-signer';
+import { findPrimaryType } from './helpers';
 
 export type Web3UnpromiEvent = Pick<PromiEvent<Contract>, 'on' | 'once'>;
 
@@ -152,36 +152,4 @@ function isProviderWithSendMethod<T extends provider>(
   provider: T
 ): provider is T & Required<Pick<AbstractProvider, 'send'>> {
   return !!provider && typeof provider === 'object' && 'send' in provider;
-}
-
-// regex from @ethersproject/hash TypedDataEncoder.constructor
-// may be overly strict, but reliable
-const baseTypeRegex = /^([^\x5b]*)(\x5b|$)/;
-
-function findPrimaryType(types: Record<string, TypedDataField[]>): string {
-  const candidates = Object.keys(types);
-  const candidatesSet = new Set(candidates);
-
-  candidates.forEach((candidate) => {
-    const typedDataFields = types[candidate];
-    if (!typedDataFields) return;
-
-    typedDataFields.forEach(({ type }) => {
-      // Get the base type (drop any array specifiers)
-      const baseType = type.match(baseTypeRegex)?.[1];
-      if (!baseType) return;
-
-      // if type was referred to as a child of another type, it can't be the primaryType
-      candidatesSet.delete(baseType);
-    });
-  });
-
-  const [primaryType] = Array.from(candidatesSet);
-
-  assert(
-    primaryType,
-    `No primary type found in SignableTypedData types, ${JSON.stringify(types)}`
-  );
-
-  return primaryType;
 }
